move static method to common util class, simplified regex
