+++ jabkit/src/main/java/org/jabref/toolkit/commands/GenerateBibFromAux.java
@@ -9,6 +9,7 @@ import org.jabref.logic.auxparser.AuxParser;
 import org.jabref.logic.auxparser.AuxParserResult;
 import org.jabref.logic.auxparser.AuxParserStatisticsProvider;
 import org.jabref.logic.auxparser.DefaultAuxParser;
+import org.jabref.logic.cleanup.FieldFormatterCleanupMapper;
 import org.jabref.logic.importer.ParserResult;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.database.BibDatabase;
@@ -80,7 +81,7 @@ class GenerateBibFromAux implements Runnable {
             return;
         }
 
-        Convert.applyFormatters(fieldFormatters, subDatabase.getEntries());
+        FieldFormatterCleanupMapper.applyFormatters(fieldFormatters, subDatabase.getEntries());
 
         if (outputFile == null) {
             System.out.println(subDatabase.getEntries().stream()
